Refactor actions.
