+++ example/src/redux/actions/storage.js
